# device-service/src/main/java/ru/yandex/practicum/device/dto/DeviceCommandDto.java
package ru.yandex.practicum.device.dto;

import lombok.Data;

@Data
public class DeviceCommandDto {

    private DeviceCommandTypes commandType;
    private Double value;

    public enum DeviceCommandTypes {
        SET_TEMPERATURE
    }
}