# telemetry-service/src/main/java/ru/yandex/practicum/telemetry/entity/Telemetry.java
package ru.yandex.practicum.telemetry.entity;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Data;

@Entity
@Table(name = "telemetry")
@Data
public class Telemetry {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private long deviceId;

    @Column(nullable = false)
    private double currentTemperature;
}